Move LiveServer to tests/ui/live_server.py
